Add overview.md
The overview contains:
- Why we're working on SourceCred
- How SourceCred works
- The current status
- How to get involved
- Some FAQs

Once this merges, I'm planning to use it as a replacement for the
current content on sourcecred.io.

I made a decision not to invoke the words "node" and "edge", prefering
"contribution" and "relationship", so as to make it more accessible for
people who aren't familiar with graphs. I welcome feedback on this
decision (and on the whole document, of course).

Test plan: Prior to sending the pull request, I will ensure that every
link works.